django/practice_project/src/requirements/base.txt
celery==5.4.0
django-extensions==3.2.3
django-redis==5.4.0
django-soft-delete==1.0.15
Django==5.1.3
djangorestframework-simplejwt>=5.2.2
djangorestframework>=3.14.0
drf-extensions==0.7.1
drf-nested-routers==0.94.1
psycopg2-binary==2.9.6
redis==5.2.0
elasticsearch==8.16.0
elasticsearch-dsl==8.16.0
django-elasticsearch-dsl==8.0
django_elasticsearch_dsl_drf==0.22.5